Move async TCP examples to separate files
